Changing Duration value of SplineWalker from float to int and enable serialization and deserialization.

+++ MonoGame.SplineFlower.Samples/Car.cs
@@ -30,7 +30,7 @@ namespace MonoGame.SplineFlower.Samples
             _CarDrive = Content.Load<SoundEffect>(@"Audio/cardrive");
         }
 
-        public override void CreateSplineWalker(BezierSpline spline, SplineWalkerMode mode, float duration, bool canTriggerEvents = true, bool autoStart = true)
+        public override void CreateSplineWalker(BezierSpline spline, SplineWalkerMode mode, int duration, bool canTriggerEvents = true, bool autoStart = true)
         {
             base.CreateSplineWalker(spline, mode, duration, canTriggerEvents, autoStart);
         }
